+++ lib/view/screens/home_screen.dart
@@ -1,3 +1,4 @@
+import 'package:asunnyday/view/widgets/search_widget.dart';
 import 'package:asunnyday/view/widgets/text_widget.dart';
 import 'package:asunnyday/view_model/internationalization/app_localizations.dart';
 import 'package:flutter/material.dart';
@@ -12,16 +13,24 @@ class HomeScreen extends StatefulWidget {
 class _HomeScreenState extends State<HomeScreen> {
   @override
   void initState() {
-
     super.initState();
   }
+
+  //Testing search widget
+  final TextEditingController _controller = TextEditingController();
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: Column(
         mainAxisAlignment: MainAxisAlignment.center,
-        children: [TextWidget( text: AppLocalizations.of(context).translate("welcome"))],
+        children: [
+          SearchWidget(
+            onCitySelected: () {},
+            typeAheadController: _controller,
+            focusNode: FocusNode(),
+          )
+        ],
       ),
     );
   }